#669 Making 'haxXYZ' the default presenceChecker
